Python 3.6.1 Django 1.11 기준으로 작성 함
Facebook Group 및 한빛미디어 홈페이지에 등록 용.

+++ photo/fields.py
@@ -26,7 +26,7 @@ class ThumbnailImageFieldFile(ImageFieldFile):
         img.thumbnail(size, Image.ANTIALIAS)
         background = Image.new('RGBA', size, (255, 255, 255, 0))
         background.paste(
-            img, ((size[0] - img.size[0]) / 2, (size[1] - img.size[1]) / 2))
+            img, ( int((size[0] - img.size[0]) / 2), int((size[1] - img.size[1]) / 2) ) )
         background.save(self.thumb_path, 'JPEG')
 
     def delete(self, save=True):
